requirements.txt
beautifulsoup4==4.11.1
Django==4.0
fastapi==0.79.1
fuzzywuzzy==0.18.0
Pillow==9.2.0
requests==2.26.0
spotipy==2.16.1
starlette==0.19.1
